+++ features/admin.users.v1/components/edit-user.tsx
@@ -16,6 +16,7 @@
  * under the License.
  */
 
+import useAuthorization from "@wso2is/admin.authorization.v1/hooks/use-authorization";
 import useUIConfig from "@wso2is/admin.core.v1/hooks/use-ui-configs";
 import { FeatureConfigInterface } from "@wso2is/admin.core.v1/models";
 import { AppState, store } from "@wso2is/admin.core.v1/store";
@@ -90,14 +91,19 @@ export const EditUser: FunctionComponent<EditUserPropsInterface> = (
 
     const { t } = useTranslation();
     const { activeTab, updateActiveTab } = useUserManagement();
+    const { legacyAuthzRuntime }  = useAuthorization();
     const dispatch: Dispatch = useDispatch();
-    const { isSuperOrganization } = useGetCurrentOrganizationType();
+    const { isSuperOrganization, isSubOrganization } = useGetCurrentOrganizationType();
     const { UIConfig } = useUIConfig();
 
     const allowedScopes: string = useSelector((state: AppState) => state?.auth?.allowedScopes);
     const isGroupAndRoleSeparationEnabled: boolean = useSelector(
         (state: AppState) => state?.config?.ui?.isGroupAndRoleSeparationEnabled);
     const roleV1Enabled: boolean = UIConfig?.legacyMode?.rolesV1;
+
+    const userRolesDisabledFeatures: string[] = useSelector((state: AppState) => {
+        return state.config.ui.features?.users?.disabledFeatures;
+    });
 
     const [ isReadOnly, setReadOnly ] = useState<boolean>(false);
     const [ isSuperAdmin, setIsSuperAdmin ] = useState<boolean>(false);
@@ -232,29 +238,37 @@ export const EditUser: FunctionComponent<EditUserPropsInterface> = (
                 </ResourceTab.Pane>
             )
         },
-        isSuperOrganization() && roleV1Enabled ? {
-            menuItem: t("users:editUser.tab.menuItems.2"),
-            render: () => (
-                <ResourceTab.Pane controlledSegmentation attached={ false }>
-                    <UserRolesV1List
-                        isGroupAndRoleSeparationEnabled={ isGroupAndRoleSeparationEnabled }
-                        onAlertFired={ handleAlerts }
-                        user={ user }
-                        handleUserUpdate={ handleUserUpdate }
-                        isReadOnly={ isReadOnly }
-                    />
-                </ResourceTab.Pane>
-            )
-        } : null,
+        !userRolesDisabledFeatures?.includes(UserManagementConstants.FEATURE_DICTIONARY.get("USER_ROLES"))
+        && !isSubOrganization()
+        && !legacyAuthzRuntime
+        && roleV1Enabled
+            ? {
+                menuItem: t("users:editUser.tab.menuItems.2"),
+                render: () => (
+                    <ResourceTab.Pane controlledSegmentation attached={ false }>
+                        <UserRolesV1List
+                            isGroupAndRoleSeparationEnabled={ isGroupAndRoleSeparationEnabled }
+                            onAlertFired={ handleAlerts }
+                            user={ user }
+                            handleUserUpdate={ handleUserUpdate }
+                            isReadOnly={ isReadOnly }
+                        />
+                    </ResourceTab.Pane>
+                )
+            } : null,
         // ToDo - Enabled only for root organizations as BE doesn't have full SCIM support for organizations yet
-        isSuperOrganization() && !roleV1Enabled ? {
-            menuItem: t("users:editUser.tab.menuItems.2"),
-            render: () => (
-                <ResourceTab.Pane controlledSegmentation attached={ false }>
-                    <UserRolesList user={ user } />
-                </ResourceTab.Pane>
-            )
-        } : null,
+        !userRolesDisabledFeatures?.includes(UserManagementConstants.FEATURE_DICTIONARY.get("USER_ROLES"))
+        && !isSubOrganization()
+        && !legacyAuthzRuntime
+        && !roleV1Enabled
+            ? {
+                menuItem: t("users:editUser.tab.menuItems.2"),
+                render: () => (
+                    <ResourceTab.Pane controlledSegmentation attached={ false }>
+                        <UserRolesList user={ user } />
+                    </ResourceTab.Pane>
+                )
+            } : null,
         {
             menuItem: t("users:editUser.tab.menuItems.3"),
             render: () => (