+++ apps/myaccount/CHANGELOG.md
@@ -1,5 +1,173 @@
 # @wso2is/myaccount
 
+## 2.7.20
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wso2is/admin.organizations.v1@2.20.23
+  - @wso2is/common.branding.v1@2.20.23
+
+## 2.7.19
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wso2is/admin.organizations.v1@2.20.22
+  - @wso2is/common.branding.v1@2.20.22
+
+## 2.7.18
+
+### Patch Changes
+
+- Updated dependencies [[`16f16d3ed494a488bf5c0cda370371a047c0de3f`](https://github.com/wso2/identity-apps/commit/16f16d3ed494a488bf5c0cda370371a047c0de3f)]:
+  - @wso2is/admin.organizations.v1@2.20.21
+  - @wso2is/common.branding.v1@2.20.21
+
+## 2.7.17
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wso2is/admin.organizations.v1@2.20.20
+  - @wso2is/common.branding.v1@2.20.20
+
+## 2.7.16
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wso2is/admin.organizations.v1@2.20.19
+  - @wso2is/common.branding.v1@2.20.19
+
+## 2.7.15
+
+### Patch Changes
+
+- Updated dependencies [[`2ed2890039a51739721918c224da1b0ffc51d7d0`](https://github.com/wso2/identity-apps/commit/2ed2890039a51739721918c224da1b0ffc51d7d0)]:
+  - @wso2is/i18n@2.3.8
+  - @wso2is/admin.organizations.v1@2.20.18
+  - @wso2is/common.branding.v1@2.20.18
+
+## 2.7.14
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wso2is/admin.organizations.v1@2.20.17
+  - @wso2is/common.branding.v1@2.20.17
+
+## 2.7.13
+
+### Patch Changes
+
+- Updated dependencies [[`3777b2553280fd18972e0fc728facfd18a55e160`](https://github.com/wso2/identity-apps/commit/3777b2553280fd18972e0fc728facfd18a55e160), [`5906e411dd31ceb41711159eddfdea025781288d`](https://github.com/wso2/identity-apps/commit/5906e411dd31ceb41711159eddfdea025781288d)]:
+  - @wso2is/admin.organizations.v1@2.20.16
+  - @wso2is/i18n@2.3.7
+  - @wso2is/common.branding.v1@2.20.16
+
+## 2.7.12
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wso2is/admin.organizations.v1@2.20.15
+  - @wso2is/common.branding.v1@2.20.15
+
+## 2.7.11
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wso2is/admin.organizations.v1@2.20.14
+  - @wso2is/common.branding.v1@2.20.14
+
+## 2.7.10
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wso2is/admin.organizations.v1@2.20.13
+  - @wso2is/common.branding.v1@2.20.13
+
+## 2.7.9
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wso2is/admin.organizations.v1@2.20.12
+  - @wso2is/common.branding.v1@2.20.12
+
+## 2.7.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wso2is/admin.organizations.v1@2.20.11
+  - @wso2is/common.branding.v1@2.20.11
+
+## 2.7.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wso2is/admin.organizations.v1@2.20.10
+  - @wso2is/common.branding.v1@2.20.10
+
+## 2.7.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wso2is/admin.organizations.v1@2.20.9
+  - @wso2is/common.branding.v1@2.20.9
+
+## 2.7.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wso2is/admin.organizations.v1@2.20.8
+  - @wso2is/common.branding.v1@2.20.8
+
+## 2.7.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wso2is/admin.organizations.v1@2.20.7
+  - @wso2is/common.branding.v1@2.20.7
+
+## 2.7.3
+
+### Patch Changes
+
+- Updated dependencies [[`e01cad085da65461dce7337ef71a1ace2fbfff85`](https://github.com/wso2/identity-apps/commit/e01cad085da65461dce7337ef71a1ace2fbfff85)]:
+  - @wso2is/i18n@2.3.6
+  - @wso2is/admin.organizations.v1@2.20.6
+  - @wso2is/common.branding.v1@2.20.6
+
+## 2.7.2
+
+### Patch Changes
+
+- Updated dependencies [[`12c6333330c4e22d30f5644800dc9b57b4ca4471`](https://github.com/wso2/identity-apps/commit/12c6333330c4e22d30f5644800dc9b57b4ca4471)]:
+  - @wso2is/i18n@2.3.5
+  - @wso2is/admin.organizations.v1@2.20.5
+  - @wso2is/common.branding.v1@2.20.5
+
+## 2.7.1
+
+### Patch Changes
+
+- [#6352](https://github.com/wso2/identity-apps/pull/6352) [`d615c10f5fd308dbfc8b5e14149dd29fcec1b99e`](https://github.com/wso2/identity-apps/commit/d615c10f5fd308dbfc8b5e14149dd29fcec1b99e) Thanks [@savindi7](https://github.com/savindi7)! - Fix header text wrap issue in resource card.
+
+- Updated dependencies [[`d615c10f5fd308dbfc8b5e14149dd29fcec1b99e`](https://github.com/wso2/identity-apps/commit/d615c10f5fd308dbfc8b5e14149dd29fcec1b99e)]:
+  - @wso2is/theme@2.0.85
+  - @wso2is/react-components@2.2.7
+  - @wso2is/admin.organizations.v1@2.20.4
+  - @wso2is/common.branding.v1@2.20.4
+
 ## 2.7.0
 
 ### Minor Changes